split tests into files
