+++ backend/websocket/test_connection.py
@@ -1,0 +1,62 @@
+from channels.testing import WebsocketCommunicator
+
+from backend.routings import application
+from .consumer import WebsocketConsumer
+from .tests import TestCaseWithCache, async_test
+from api.models import User
+
+
+class WebsocketConnectionTestCase(TestCaseWithCache):
+    @async_test
+    async def test_unauthenticated(self):
+        communicator = WebsocketCommunicator(application, '/ws/party/')
+        resp = await communicator.connect(1)
+        self.assertTupleEqual(resp, (False, 4000))
+
+    @async_test
+    async def test_connect(self):
+        user = User.objects.create_user(
+            email='ferris@rustacean.org',
+            password='iluvrust',
+            username='ferris',
+        )
+        self.client.login(email=user.email, password='iluvrust')
+
+        communicator = WebsocketCommunicator(WebsocketConsumer, '/',)
+        communicator.scope['user'] = user
+
+        connected, _ = await communicator.connect()
+        self.assertTrue(connected)
+        self.assertEqual(communicator.instance.scope['user'], user)
+
+        await communicator.disconnect()
+
+    @async_test
+    async def test_unauthenticated_during_connection(self):
+        from django.contrib.auth.models import AnonymousUser
+
+        user = User.objects.create_user(
+            email='ferris@rustacean.org',
+            password='iluvrust',
+            username='ferris',
+        )
+        self.client.login(email=user.email, password='iluvrust')
+
+        communicator = WebsocketCommunicator(WebsocketConsumer, '/',)
+        communicator.scope['user'] = user
+
+        await communicator.connect()
+        await communicator.receive_json_from(1)
+
+        self.client.logout()
+        communicator.scope['user'] = AnonymousUser()
+
+        await communicator.send_json_to({
+            'command': 'party.join',
+            'foo': 'bar',
+        })
+        resp = await communicator.receive_output(1)
+        self.assertEqual(resp['type'], 'websocket.close')
+        self.assertEqual(resp['code'], 4010)
+
+        await communicator.disconnect()
